+++ src/misc/SequentialAnimation.h
@@ -16,7 +16,7 @@ class SequentialAnimation : public AnimationGroup
 
 public:
     virtual void start();
-    virtual void update(float dt);
+    virtual void update(double dt);
 
 private:
     BaseAnimation* nextAnimation();
